Rename AccountDB folders
