+++ ironfish/src/rpc/routes/accounts/types.ts
@@ -1,8 +1,8 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-import { DecryptedNoteValue } from '../../../wallet/database/decryptedNoteValue'
-import { TransactionValue } from '../../../wallet/database/transactionValue'
+import { DecryptedNoteValue } from '../../../wallet/walletdb/decryptedNoteValue'
+import { TransactionValue } from '../../../wallet/walletdb/transactionValue'
 
 export type RpcAccountTransaction = {
   hash: string
