+++ ironfish/src/wallet/account.ts
@@ -10,10 +10,10 @@ import { DatabaseKeyRange, IDatabaseTransaction } from '../storage'
 import { StorageUtils } from '../storage/database/utils'
 import { BufferUtils } from '../utils'
 import { DecryptedNote } from '../workerPool/tasks/decryptNotes'
-import { AccountsDB } from './database/accountsdb'
-import { AccountValue } from './database/accountValue'
-import { DecryptedNoteValue } from './database/decryptedNoteValue'
-import { TransactionValue } from './database/transactionValue'
+import { AccountsDB } from './walletdb/walletdb'
+import { AccountValue } from './walletdb/accountValue'
+import { DecryptedNoteValue } from './walletdb/decryptedNoteValue'
+import { TransactionValue } from './walletdb/transactionValue'
 import { SyncTransactionParams } from './wallet'
 
 export const ACCOUNT_KEY_LENGTH = 32
