+++ ironfish/src/wallet/walletdb/__fixtures__/decryptedNoteValue.test.ts.fixture
