+++ ironfish/src/wallet/walletdb/__fixtures__/transactionValue.test.ts.fixture
