+++ ironfish/src/wallet/walletdb/decryptedNoteValue.test.ts
