+++ ironfish/src/wallet/walletdb/decryptedNoteValue.ts
