+++ ironfish/src/wallet/walletdb/transactionValue.ts
